refactor: simplify with primary ctors
