+++ src/WHyLL/ByteWarp/ByteWarpEnvelope.cs
