--- src/WHyLL/IOptional.cs
@@ -1,10 +1,0 @@
-namespace Xemo;
-
-public interface IOptional<TValue>
-{
-    bool Has();
-
-    IOptional<TValue> IfHas(Action<TValue> action);
-    IOptional<TValue> IfNot(Action action);
-    TValue Out();
-}
