+++ src/WHyLL/Message/AsyncMessage.cs
@@ -1,6 +1,5 @@
 ﻿using Tonga;
 using WHyLL.Warp;
-using WHyLL.Warp;
 
 namespace WHyLL.Message
 {
