+++ src/WHyLL/Warp/BodyAsJsonArray.cs
@@ -1,5 +1,6 @@
 using Newtonsoft.Json.Linq;
-using WHyLL.Warp;
+
+namespace WHyLL.Warp;
 
 /// <summary>
 /// Body as json array.
