+++ src/WHyLL/Warp/FirstHeader.cs
@@ -1,13 +1,9 @@
-﻿using Tonga.Scalar;
+﻿namespace WHyLL.Warp;
 
-namespace WHyLL.Warp
-{
-    /// <summary>
-    /// First header with the given name.
-    /// </summary>
-    public sealed class FirstHeader(string name) : WarpEnvelope<string>(
-        new FirstHeaderAs<string>(name, header => header)
-    )
-    { }
-}
+/// <summary>
+/// First header with the given name.
+/// </summary>
+public sealed class FirstHeader(string name) : WarpEnvelope<string>(
+    new FirstHeaderAs<string>(name, header => header)
+);
