+++ src/WHyLL/Warp/FirstHeaderAsInt.cs
@@ -1,11 +1,8 @@
-﻿namespace WHyLL.Warp
-{
-    /// <summary>
-    /// First header with the given name as int.
-    /// </summary>
-    public sealed class FirstHeaderAsInt(string name) : WarpEnvelope<int>(
-        new FirstHeaderAs<int>(name, Convert.ToInt32)
-    )
-    { }
-}
+﻿namespace WHyLL.Warp;
 
+/// <summary>
+/// First header with the given name as int.
+/// </summary>
+public sealed class FirstHeaderAsInt(string name) : WarpEnvelope<int>(
+    new FirstHeaderAs<int>(name, Convert.ToInt32)
+);
