+++ src/WHyLL/Warp/FirstLine.cs
@@ -1,13 +1,9 @@
-﻿using Tonga;
+﻿namespace WHyLL.Warp;
 
-namespace WHyLL.Warp
-{
-    /// <summary>
-    /// Render the first line of a request (request-line) or response (status-line)
-    /// </summary>
-	public sealed class FirstLine() : WarpEnvelope<string>(
-	    new FirstLineAs<string>(line => line)
-	)
-	{ }
-}
+/// <summary>
+/// Render the first line of a request (request-line) or response (status-line)
+/// </summary>
+public sealed class FirstLine() : WarpEnvelope<string>(
+    new FirstLineAs<string>(line => line)
+);
