+++ tests/Test.WHyLL/Warp/BodyAsJsonArrayTest.cs
@@ -1,5 +1,6 @@
 ﻿using Tonga.IO;
 using WHyLL.Message;
+using WHyLL.Warp;
 using Xunit;
 
 namespace Test.WHyLL.Warp
